+++ buildSrc/src/main/kotlin/Versions.kt
@@ -4,10 +4,10 @@ object Versions {
     const val androidTargetSdk = 34
     const val binaryCompatibilityValidatorGradlePlugin = "0.16.3"
     const val jvmTarget = "1.8"
-    const val kotlin = "2.0.10"
-    const val kotlinCoroutines = "1.9.0-RC"
+    const val kotlin = "2.0.20"
+    const val kotlinCoroutines = "1.9.0-RC.2"
     const val kotlinIo = "0.5.3"
-    const val kotlinSerialization = "1.7.1"
+    const val kotlinSerialization = "1.7.2"
     const val robolectric = "11-robolectric-6757853"
     const val osDetectorGradlePlugin = "1.7.3"
     const val protoc = "4.28.0"